+++ testing/CHANGELOG.md
@@ -1,5 +1,14 @@
 # testing
 
+## 0.0.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @bs-core/shell@1.12.0
+  - @bs-plugins/jira@1.0.17
+  - @bs-plugins/template@0.0.17
+
 ## 0.0.15
 
 ### Patch Changes